requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
attrs==25.4.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
deprecation==2.1.0
dnspython==2.8.0
eventlet==0.40.3
eyeD3==0.9.8
filetype==1.2.0
Flask==3.1.2
Flask-SocketIO==5.5.1
frozenlist==1.8.0
gevent==25.9.1
gevent-websocket==0.10.1
greenlet==3.2.4
h11==0.16.0
h2==4.3.0
hpack==4.1.0
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
marisa-trie==1.3.1
MarkupSafe==3.0.3
multidict==6.7.0
mutagen==1.47.0
packaging==25.0
pillow==12.0.0
priority==2.0.0
propcache==0.4.1
psutil==7.1.3
pycparser==2.23
python-engineio==4.12.3
python-socketio==5.14.3
python-vlc==3.0.21203
pywin32==311
Quart==0.20.0
simple-websocket==1.1.0
urllib3==2.5.0
uvicorn==0.38.0
websocket-client==1.9.0
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.22.0
zope.event==6.1
zope.interface==8.1